+++ packages/create-codeceptjs-bdd-tests/acceptance/specs/login.mocha.spec.ts
@@ -1,3 +1,5 @@
+const softAssert = require('soft-assert');
+
 Feature('Login Tests (mocha)');
 
 Before(({ I }) => {
@@ -5,39 +7,7 @@ Before(({ I }) => {
 });
 
 Scenario('Fred logs in successfully', ({ I }) => {
-    I.amOnPage('/gkushang/codeceptjs-bdd');
+    I.amOnPage('https://github.com');
 })
     .tag('@login_mocha')
     .tag('smoke');
-
-Scenario('Fred intercept request for playwright @intercept', async ({ I }) => {
-    if (await I.checkIfRunningOnPlaywright()) {
-        // Set Mock Data
-        const mockResponseObject = [
-            {
-                id: 1,
-                title: 'How to Intercept a Response in Playwright',
-                author: 'CodeceptBDD',
-                genre: 'business',
-                price: '100.00',
-                rating: '★★★★★',
-                stock: 65535,
-            },
-        ];
-
-        // Intercept Request
-        I.usePlaywrightTo('do intercept', async ({ page }) => {
-            await page.route('https://danube-webshop.herokuapp.com/api/books', (route) =>
-                route.fulfill({
-                    contentType: 'application/json',
-                    body: JSON.stringify(mockResponseObject),
-                })
-            );
-        });
-
-        // Validate
-        I.amOnPage('https://danube-webshop.herokuapp.com/');
-        I.seeTextEquals('How to Intercept a Response in Playwright', { css: '.preview-title' });
-        I.seeTextEquals('CodeceptBDD', { css: '.preview-author' });
-    }
-});